Fjernet NuGetpakke, Endret konstruktører i Dokument.
