--- SikkerDigitalPost.Net.KlientDemo/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="SikkerDigitalPostKlientbibliotek" version="0.1" targetFramework="net45" />
-</packages>
